Releases/1.0.x

+++ shop/views/auth.py
@@ -13,25 +13,26 @@ from rest_framework.views import APIView
 from rest_framework.exceptions import ValidationError
 from rest_framework.generics import GenericAPIView
 from rest_framework.permissions import AllowAny
-from rest_framework.renderers import TemplateHTMLRenderer, JSONRenderer, BrowsableAPIRenderer
+from rest_framework.renderers import JSONRenderer, BrowsableAPIRenderer
 from rest_framework.response import Response
 from rest_auth.views import LoginView as OriginalLoginView, PasswordChangeView as OriginalPasswordChangeView
 
 from shop.models.cart import CartModel
 from shop.models.customer import CustomerModel
-from shop.rest.auth import PasswordResetSerializer, PasswordResetConfirmSerializer
+from shop.rest.renderers import CMSPageRenderer
+from shop.serializers.auth import PasswordResetRequestSerializer, PasswordResetConfirmSerializer
 from shop.signals import email_queued
 
 
 class AuthFormsView(GenericAPIView):
     """
-    Generic view to handle authetication related forms such as user registration
+    Generic view to handle authentication related forms such as user registration
     """
     serializer_class = None
     form_class = None
 
     def post(self, request, *args, **kwargs):
-        if request.customer.is_visitor():
+        if request.customer.is_visitor:
             customer = CustomerModel.objects.get_or_create_from_request(request)
         else:
             customer = request.customer
@@ -50,6 +51,8 @@ class AuthFormsView(GenericAPIView):
 
 
 class LoginView(OriginalLoginView):
+    form_name = 'login_form'
+
     def login(self):
         """
         Logs in as the given user, and moves the items from the current to the new cart.
@@ -58,18 +61,22 @@ class LoginView(OriginalLoginView):
             anonymous_cart = CartModel.objects.get_from_request(self.request)
         except CartModel.DoesNotExist:
             anonymous_cart = None
-        if self.request.customer.user.is_anonymous() or self.request.customer.is_registered():
+        if self.request.customer.user.is_anonymous or self.request.customer.is_registered:
             previous_user = None
         else:
             previous_user = self.request.customer.user
         super(LoginView, self).login()  # this rotates the session_key
+        if not self.serializer.data.get('stay_logged_in'):
+            self.request.session.set_expiry(0)  # log out when the browser is closed
         authenticated_cart = CartModel.objects.get_from_request(self.request)
         if anonymous_cart:
             # an anonymous customer logged in, now merge his current cart with a cart,
             # which previously might have been created under his account.
             authenticated_cart.merge_with(anonymous_cart)
         if previous_user and previous_user.is_active is False and previous_user != self.request.user:
-            previous_user.delete()  # keep the database clean and remove this anonymous entity
+            # keep the database clean and remove this anonymous entity
+            if previous_user.customer.orders.count() == 0:
+                previous_user.delete()
 
     def post(self, request, *args, **kwargs):
         self.request = request
@@ -79,7 +86,7 @@ class LoginView(OriginalLoginView):
             self.login()
             return self.get_response()
 
-        exc = ValidationError({'login_form': self.serializer.errors})
+        exc = ValidationError({self.form_name: self.serializer.errors})
         response = self.handle_exception(exc)
         self.response = self.finalize_response(request, response, *args, **kwargs)
         return self.response
@@ -90,6 +97,7 @@ class LogoutView(APIView):
     Calls Django logout method and delete the auth Token assigned to the current User object.
     """
     permission_classes = (AllowAny,)
+    form_name = 'logout_form'
 
     def post(self, request):
         try:
@@ -98,38 +106,43 @@ class LogoutView(APIView):
             pass
         logout(request)
         request.user = AnonymousUser()
-        response_data = {'logout_form': {'success_message': _("Successfully logged out.")}}
+        response_data = {self.form_name: {'success_message': _("Successfully logged out.")}}
         return Response(response_data)
 
 
 class PasswordChangeView(OriginalPasswordChangeView):
+    form_name = 'password_change_form'
+
     def post(self, request, *args, **kwargs):
         form_data = request.data.get('form_data', {})
         serializer = self.get_serializer(data=form_data)
         if serializer.is_valid():
             serializer.save()
-            response_data = {'change_form': {
+            response_data = {self.form_name: {
                 'success_message': _("Password has been changed successfully."),
             }}
             return Response(response_data)
-        return Response({'change_form': serializer.errors}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+        return Response({self.form_name: serializer.errors}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
 
-class PasswordResetView(GenericAPIView):
+class PasswordResetRequestView(GenericAPIView):
     """
-    Calls Django Auth PasswordResetForm save method.
+    Calls Django Auth PasswordResetRequestForm save method.
 
     Accepts the following POST parameters: email
     Returns the success/fail message.
     """
-    serializer_class = PasswordResetSerializer
+    serializer_class = PasswordResetRequestSerializer
     permission_classes = (AllowAny,)
+    form_name = 'password_reset_request_form'
 
     def post(self, request, *args, **kwargs):
         form_data = request.data.get('form_data', {})
         serializer = self.get_serializer(data=form_data)
         if not serializer.is_valid():
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({self.form_name: serializer.errors}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+
+        # send email containing a reset link
         serializer.save()
 
         # trigger async email queue
@@ -137,33 +150,23 @@ class PasswordResetView(GenericAPIView):
 
         # Return the success message with OK HTTP status
         msg = _("Instructions on how to reset the password have been sent to '{email}'.")
-        response_data = {'reset_form': {
+        response_data = {self.form_name: {
             'success_message': msg.format(**serializer.data),
         }}
         return Response(response_data)
 
 
-class PasswordResetConfirm(GenericAPIView):
+class PasswordResetConfirmView(GenericAPIView):
     """
-    Password reset e-mail link points onto this view, which when invoked by a GET request renderes
-    a HTML page containing a password reset form. This form then can be used to reset the user's
-    password using a RESTful POST request.
-
-    Since the URL for this view is part in the email's body text, expose it to the URL patterns as:
-
-    ```
-    url(r'^password-reset-confirm/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
-        PasswordResetConfirm.as_view(), name='password_reset_confirm'),
-    ```
-
-    Accepts the following POST parameters: new_password1, new_password2
-    Returns the success/fail message.
+    Password reset e-mail link points onto a CMS page with the Page ID = 'password-reset-confirm'.
+    This page then shall render the CMS plugin as provided by the **ShopAuthenticationPlugin** using
+    the form "Confirm Password Reset".
     """
-    renderer_classes = (TemplateHTMLRenderer, JSONRenderer, BrowsableAPIRenderer)
+    renderer_classes = (CMSPageRenderer, JSONRenderer, BrowsableAPIRenderer)
     serializer_class = PasswordResetConfirmSerializer
     permission_classes = (AllowAny,)
-    template_name = 'shop/auth/password-reset-confirm.html'
     token_generator = default_token_generator
+    form_name = 'password_reset_confirm_form'
 
     def get(self, request, uidb64=None, token=None):
         data = {'uid': uidb64, 'token': token}
@@ -180,11 +183,18 @@ class PasswordResetConfirm(GenericAPIView):
         })
 
     def post(self, request, uidb64=None, token=None):
-        data = dict(request.data, uid=uidb64, token=token)
-        serializer = self.get_serializer(data=data)
-        if not serializer.is_valid():
-            return Response(
-                serializer.errors, status=status.HTTP_400_BAD_REQUEST
-            )
-        serializer.save()
-        return Response({"success": _("Password has been reset with the new password.")})
+        try:
+            data = dict(request.data['form_data'], uid=uidb64, token=token)
+        except (KeyError, TypeError, ValueError):
+            errors = {'non_field_errors': [_("Invalid POST data.")]}
+        else:
+            serializer = self.get_serializer(data=data)
+            if serializer.is_valid():
+                serializer.save()
+                response_data = {self.form_name: {
+                    'success_message': _("Password has been reset with the new password."),
+                }}
+                return Response(response_data)
+            else:
+                errors = serializer.errors
+        return Response({self.form_name: errors}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
